# server,heap: severely enhance performance for delete_from_heaps