As we discussed in https://github.com/ceph/dmclock/pull/67, **do_clean** might take minutes to erase the old clients if there are thousands of clients became old enough to erase, which result in some unanticipated situation such as request been blocked. so I pushed that PR(https://github.com/ceph/dmclock/pull/67) to limit number of old clients to be erased at a time, it can largely ameliorate this problem. However, it doesn't look perfect since it didn't reduce the time to erase one old client.

Recently , I dig into the code further more and find the time is mainly consumed by **delete_from_heaps** at [#L258](https://github.com/ceph/dmclock/blob/master/support/src/indirect_intrusive_heap.h#L258), and finally [#L223](https://github.com/ceph/dmclock/blob/master/support/src/indirect_intrusive_heap.h#L223).

This change include three ways:

-  use **pointer heap** to enhance performance for Iterator
-  add and use a new method **at** for class IndIntruHeap 
-  increase standard_erase_max as delete_from_heaps become more efficient 

We add some testing code to test the time consumed by **delete_from_heaps**  to erase 100 old clients at a time, the result is show as below:
1. **origin** without this change:
```
### delete_from_heaps consume time: 0.840726s, erased: 100, client_map: 12899
### delete_from_heaps consume time: 1.39886s, erased: 100, client_map: 20499
### delete_from_heaps consume time: 1.84538s, erased: 100, client_map: 31899
### delete_from_heaps consume time: 2.84816s, erased: 100, client_map: 40899
### delete_from_heaps consume time: 3.25994s, erased: 100, client_map: 50699
### delete_from_heaps consume time: 3.62733s, erased: 100, client_map: 60599
### delete_from_heaps consume time: 5.427s, erased: 100, client_map: 70099
### delete_from_heaps consume time: 5.46223s, erased: 100, client_map: 70400
```

2. with this **pointer heap** change:
```
### delete_from_heaps consume time: 0.0147157s, erased: 100, client_map: 12899
### delete_from_heaps consume time: 0.0150367s, erased: 100, client_map: 21399
### delete_from_heaps consume time: 0.0374631s, erased: 100, client_map: 31199
### delete_from_heaps consume time: 0.0375769s, erased: 100, client_map: 40999
### delete_from_heaps consume time: 0.0368521s, erased: 100, client_map: 50799
### delete_from_heaps consume time: 0.0645182s, erased: 100, client_map: 60199
```

3. with this **pointer heap** and **at** changes:
```
### delete_from_heaps consume time: 0.00179381s, erased: 100, client_map: 12399
### delete_from_heaps consume time: 0.00249408s, erased: 100, client_map: 20399
### delete_from_heaps consume time: 0.00224836s, erased: 100, client_map: 30073
### delete_from_heaps consume time: 0.00234324s, erased: 100, client_map: 40599
### delete_from_heaps consume time: 0.00492391s, erased: 100, client_map: 50099
### delete_from_heaps consume time: 0.00838497s, erased: 100, client_map: 60399
### delete_from_heaps consume time: 0.00867261s, erased: 100, client_map: 61033
```
Signed-off-by: yanjun <yan.jun8@zte.com.cn>