@Yan-waller 2019-02-27
>  My question remains, though -- why would pointers be faster than references?

```c++
      Iterator(IndIntruHeap<I, T, heap_info, C, K>& _heap, HeapIndex _index) :
	heap(_heap),
	index(_index)
      {
	// empty
      }
```
It seems that the above initialization of heap call the `IndIntruHeap` constructor, ten thousands of `data` items will be copied and do `push_back` as following.
```c++
    IndIntruHeap(const IndIntruHeap<I,T,heap_info,C,K>& other) :
      count(other.count)
    {
      for (HeapIndex i = 0; i < other.count; ++i) {
	data.push_back(other.data[i]);
      }
    }
```
If we use pointers heap, the initialization `heap(_heap)` is just copy a pointer.

Finally, I increase the standard_erase_max (from 100 to 2000, or may be some other value) based on this optimization since delete_from_heaps has became more efficient.