@Yan-waller 2019-02-28
@ivancich I think I see what you mean, and could you re-check the code?

below is the latest test result:
```
### delete_from_heaps consume time: 0.00633848s, erased: 2000, client_map: 17999
### delete_from_heaps consume time: 0.00505186s, erased: 2000, client_map: 18999
### delete_from_heaps consume time: 0.00688795s, erased: 2000, client_map: 19999
```